+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Unreleased
 
+- added: Report legacy payment ID's using the new core `memo` transaction property.
 - removed: Only bundle the package in CJS-format, and drop the ESM output.
 - removed: Stop supporting undocumented `getTransaction` options.